Updating the readme file and fixing the errors in the controllers
