+++ app/controllers/application_controller.rb
@@ -1,25 +1,7 @@
+require_relative('../../services/authentication_helper.rb')
+
 class ApplicationController < ActionController::API
-  def encode_token(payload)
-    JWT.encode(payload, 'sgubp')
-  end
-
-  def auth_header
-    # { Authorization: 'Bearer <token>' }
-    request.headers['Authorization']
-  end
-
-  def decoded_token
-    return unless auth_header
-
-    token = auth_header.split(' ')[1]
-
-    begin
-      JWT.decode(token, 'sgubp', true, algorithm: 'HS256')
-    rescue JWT::DecodeError
-      nil
-    end
-  end
-
+  
   def logged_in_user
     return unless decoded_token
 
@@ -32,6 +14,6 @@ class ApplicationController < ActionController::API
   end
 
   def authorized
-    render json: { message: 'Please log in', status: 404 }, status: :unauthorized unless logged_in?
+    render json: { message: 'Please log in' }, status: :unauthorized unless logged_in?
   end
 end
