+++ app/models/favourite.rb
@@ -1,5 +1,5 @@
 class Favourite < ApplicationRecord
   belongs_to :house
   belongs_to :user
-  validates_uniqueness_of :house
+  validates :house,uniqueness: {scope: :user}
 end
